refactor(docs): implement lazy-loading
